fix disable scroll hook
