+++ docs/fds-components/index.html
@@ -16,4 +16,4 @@
   }</script><style>#root[hidden],
       #docs-root[hidden] {
         display: none !important;
-      }</style></head><body><div id="root"></div><div id="docs-root"></div><script>window['DOCS_MODE'] = false;</script><script src="runtime~main.1082140af9db1185c9a2.bundle.js"></script><script src="vendors~main.505662079ee9f5908cde.bundle.js"></script><script src="main.881086631876eb6cc63a.bundle.js"></script></body></html>
+      }</style></head><body><div id="root"></div><div id="docs-root"></div><script>window['DOCS_MODE'] = false;</script><script src="runtime~main.1082140af9db1185c9a2.bundle.js"></script><script src="vendors~main.eaba69e0d7cd5b0623f0.bundle.js"></script><script src="main.d5fe5a26f0d6be214e74.bundle.js"></script></body></html>
