+++ docs/fds-styles/base-styles/base-styles.full.css
@@ -1,7 +1,7 @@
 /* Custom properties */
 /**
  * Do not edit directly
- * Generated on Mon, 18 May 2020 19:17:44 GMT
+ * Generated on Wed, 10 Jun 2020 20:59:11 GMT
  */
 :root {
   --border-color-input: rgba(64, 64, 64, 0.24);
@@ -78,7 +78,7 @@
 }
 /**
 * DO NOT EDIT
-* Generated by fds-dictionary on Mon May 18 2020 15:17:44 GMT-0400 (Eastern Daylight Time)
+* Generated by fds-dictionary on Wed Jun 10 2020 16:59:11 GMT-0400 (Eastern Daylight Time)
 * github.com/cbinsights/form-design-system/
 */
 /**
