+++ docs/fds-styles/index.html
@@ -1437,7 +1437,7 @@ of clamping text to one line.</p>
         </main>
       </div>
       <footer class="footer">
-        Last modified Monday, 18 May 2020 15:17
+        Last modified Wednesday, 10 June 2020 16:59
       </footer>
     </div>
     <script src="prism.js?871234"></script>
