+++ src/components/Popover/util.js
@@ -1,0 +1,15 @@
+/* eslint-disable import/prefer-default-export */
+
+/**
+ * https://popper.js.org/popper-documentation.html#Popper.placements
+ *
+ * @param {String} position prop value
+ * @param {String} alignment prop value
+ * @returns {String} valid `placement` value for PopperJS `Popper` component
+ */
+export const getPopperPlacement = (position, alignment) => {
+  // Popperjs does not have a "center" placement variation. It centers by default.
+  // Our component explicitly accepts a `center` value for alignment.
+  const variation = !alignment || alignment === 'center' ? '' : `-${alignment}`;
+  return `${position}${variation}`;
+};
